+++ src/Command/Crawler/DisqusCrawler.php
@@ -80,49 +80,4 @@ class DisqusCrawler extends AbstractCrawler
         parent::setItem($item);
         $this->identifier = str_replace('{id}', $item->id, $this->config['identifier']);
     }
-
-
-
-
-
-//    public function getCountMost($interval = '30d', $limit = 25)
-//    {
-//        $response = $this->client->request(
-//            'GET',
-//            'https://disqus.com/api/3.0/threads/listPopular.json?api_key=' . $this->apikey . '&forum=' . $this->forum . '&interval=' . $interval . '&limit=' . $limit
-//        );
-//
-//        $posts = [];
-//
-//        if ($response->getStatusCode() == 200)
-//        {
-//            $data = json_decode($response->getBody()->getContents(), true);
-//
-//            if (!is_array())
-//            {
-//                return $posts;
-//            }
-//
-//            foreach ($data['response'] as $thread)
-//            {
-//                $identifiers = $thread['identifiers'];
-//                $identifier  = $identifiers[count($identifiers) - 1];
-//                $parts       = explode("-", $identifier);
-//                if (count($parts) == 3)
-//                { // format: news-uid-12
-//                    $uid         = intval($parts[2]);
-//                    $posts[$uid] = intval($thread['posts']);
-//                }
-//                if (count($parts) == 2)
-//                { // format: uid-12
-//                    $uid         = intval($parts[1]);
-//                    $posts[$uid] = intval($thread['posts']);
-//                }
-//            }
-//        }
-//
-//        return $posts;
-//    }
 }
-
-?>