+++ src/Command/SocialstatssyncCommand.php
@@ -22,7 +22,6 @@ use HeimrichHannot\NewsBundle\Command\Crawler\GoogleAnalyticsCrawler;
 use HeimrichHannot\NewsBundle\Command\Crawler\GooglePlusCrawler;
 use HeimrichHannot\NewsBundle\Command\Crawler\TwitterCrawler;
 use GuzzleHttp\Client;
-use Model\Collection;
 use Monolog\Logger;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -55,6 +54,11 @@ class SocialstatssyncCommand extends AbstractLockedCommand implements FrameworkA
     private $io;
 
     /**
+     * @var null|NewsModel|\Contao\Model\Collection
+     */
+    private $items = null;
+
+    /**
      * {@inheritdoc}
      */
     protected function configure()
@@ -63,7 +67,9 @@ class SocialstatssyncCommand extends AbstractLockedCommand implements FrameworkA
             ->setName('huh:news:socialstats')
             ->setDescription('Updates the database with social stats.')
             ->addOption('no-chunksize', null, null, "Set to 1 to ignore the limit set in options (means all results). Default 0.")
-            ->addOption('no-days', null, null, "Set to 1 to ignore the days settings (means there is no limit due age of the news).");
+            ->addOption('no-days', null, null, "Set to 1 to ignore the days settings (means there is no limit due age of the news).")
+            ->addOption('only-current', null, null, "Update latest articles.");
+
     }
 
     /**
@@ -94,9 +100,23 @@ class SocialstatssyncCommand extends AbstractLockedCommand implements FrameworkA
             'social_stats'              => System::getContainer()->getParameter('social_stats'),
             'ssl.certificate_authority' => false
         ]);
+
+        if ($input->getOption('only-current') == 1)
+        {
+            /**
+             * @var NewsModel $model
+             */
+            $model = $this->framework->getAdapter(NewsModel::class);
+            if ($model)
+            {
+                $this->items = $model->findPublishedFromToByPids(0, time(), $this->config['archives'], $this->config['chunksize']);
+            }
+
+        }
+
         try
         {
-//            $this->updateGoogleAnalyticsCount();
+            $this->updateGoogleAnalyticsCount();
             $this->updateFacebookCount();
             $this->updateTwitterCount();
             $this->updateGooglePlusCount();
@@ -116,20 +136,14 @@ class SocialstatssyncCommand extends AbstractLockedCommand implements FrameworkA
      */
     private function updateGoogleAnalyticsCount()
     {
-        if (!array_key_exists('google_analytics', $this->config))
-        {
-            $this->io->note("No Google Analytics config provided. Skipping...");
-            return;
-        }
-        $items = NewsModel::findByGoogleAnalyticsUpdateDate(
-            $this->config['chunksize'],
-            $this->config['days'],
-            $this->config['archives']
-        );
+        $crawlerConfig = [
+            'name' => 'Google Analytics',
+            'alias' => 'google_analytics',
+            'method' => 'findByGoogleAnalyticsUpdateDate'
+        ];
         $this->updateStats(
             new GoogleAnalyticsCrawler($this->httpClient, null, '', $this->config['google_analytics']),
-            $items,
-            'Google Analytics'
+            $crawlerConfig
         );
     }
 
@@ -138,16 +152,14 @@ class SocialstatssyncCommand extends AbstractLockedCommand implements FrameworkA
      */
     private function updateFacebookCount()
     {
-        if (!array_key_exists('facebook', $this->config))
-        {
-            $this->io->note("No Facebook config provided. Skipping...");
-            return;
-        }
-        $items = NewsModel::findByFacebookCounterUpdateDate($this->config['chunksize'], $this->config['days'], $this->config['archives']);
+        $crawlerConfig = [
+            'name' => 'Facebook',
+            'alias' => 'facebook',
+            'method' => 'findByFacebookCounterUpdateDate'
+        ];
         $this->updateStats(
             new FacebookCrawler($this->httpClient),
-            $items,
-            'Facebook'
+            $crawlerConfig
         );
     }
 
@@ -156,20 +168,14 @@ class SocialstatssyncCommand extends AbstractLockedCommand implements FrameworkA
      */
     private function updateTwitterCount()
     {
-        if (!$this->config['twitter'])
-        {
-            $this->io->note("No Twitter config provided. Skipping...");
-            return;
-        }
-        $items = NewsModel::findByTwitterCounterUpdateDate(
-            $this->config['chunksize'],
-            $this->config['days'],
-            $this->config['archives']
-        );
+        $crawlerConfig = [
+            'name' => 'Twitter',
+            'alias' => 'twitter',
+            'method' => 'findByTwitterCounterUpdateDate'
+        ];
         $this->updateStats(
             new TwitterCrawler($this->httpClient, null, $this->baseUrl, $this->config['twitter']),
-            $items,
-            'Twitter'
+            $crawlerConfig
         );
     }
 
@@ -178,16 +184,14 @@ class SocialstatssyncCommand extends AbstractLockedCommand implements FrameworkA
      */
     private function updateGooglePlusCount()
     {
-        if (!array_key_exists('google_plus', $this->config))
-        {
-            $this->io->note("No Google Plus config provided. Skipping...");
-            return;
-        }
-        $items = NewsModel::findByGooglePlusCounterUpdateDate($this->config['chunksize'], $this->config['days'], $this->config['archives']);
+        $crawlerConfig = [
+            'name' => 'Google Plus',
+            'alias' => 'google_plus',
+            'method' => 'findByGooglePlusCounterUpdateDate'
+        ];
         $this->updateStats(
             new GooglePlusCrawler($this->httpClient),
-            $items,
-            'Google Plus'
+            $crawlerConfig
         );
     }
 
@@ -196,26 +200,43 @@ class SocialstatssyncCommand extends AbstractLockedCommand implements FrameworkA
      */
     private function updateDisqusCount()
     {
-        if (!$this->config['disqus'])
-        {
-            $this->io->note("No Disqus config provided. Skipping...");
-            return;
-        }
-        $items = NewsModel::findByDisqusCounterUpdateDate($this->config['chunksize'], $this->config['days'], $this->config['archives']);
+        $crawlerConfig = [
+            'name' => 'Disqus',
+            'alias' => 'disqus',
+            'method' => 'findByDisqusCounterUpdateDate'
+        ];
         $this->updateStats(
             new DisqusCrawler($this->httpClient, null, $this->baseUrl, $this->config['disqus']),
-            $items,
-            'Disqus'
+            $crawlerConfig
         );
     }
 
     /**
      * @param AbstractCrawler $crawler
-     * @param NewsModel|Collection $items
+     * @param array $crawlerConfig
+
      */
-    private function updateStats($crawler, $items, $provider)
+    private function updateStats(AbstractCrawler $crawler, array $crawlerConfig)
     {
-        $this->io->section("Retriving $provider counts");
+        if (!array_key_exists($crawlerConfig['alias'], $this->config))
+        {
+            $this->io->note("No ".$crawlerConfig['name']." config provided. Skipping...");
+            return;
+        }
+        $this->io->section("Retriving ".$crawlerConfig["name"]." counts");
+        if (!$this->items)
+        {
+            $method = $crawlerConfig["method"];
+            $items = NewsModel::$method(
+                $this->config['chunksize'],
+                $this->config['days'],
+                $this->config['archives']
+            );
+        }
+        else {
+            $items = $this->items;
+        }
+
         foreach ($items as $item)
         {
             $this->io->text('Updating news article ' . $item->id . ' (' . $item->headline . ')');
@@ -225,7 +246,7 @@ class SocialstatssyncCommand extends AbstractLockedCommand implements FrameworkA
             if (is_array($count))
             {
                 $this->io->note('Error: ' . $count['message']);
-                $this->logger->addNotice($provider . ': ' . $count['message']);
+                $this->logger->addNotice($crawlerConfig['name'] . ': ' . $count['message']);
                 if ($count['code'] == AbstractCrawler::ERROR_BREAKING)
                 {
                     $this->io->note("Stopping updating stats for current provider.");
@@ -236,7 +257,7 @@ class SocialstatssyncCommand extends AbstractLockedCommand implements FrameworkA
                 }
             }
             $crawler->updateItem();
-            $this->io->text('Found ' . $count . ' shares for ' . $provider . '.');
+            $this->io->text('Found ' . $count . ' shares for ' . $crawlerConfig['name'] . '.');
         }
     }
 }